+++ src/models/classes/paymentMethod/AmexPaymentMethod.php
@@ -100,6 +100,10 @@ class AmexPaymentMethod extends PaymentMethod
         }
 
         $resource_details = parent::getResourceDetail($resource_id);
+        if (empty($resource_details)) {
+            return $resource_details;
+        }
+
         $translation = $this->dependencies
             ->getPlugin()
             ->getTranslationClass()