+++ changelog.md
@@ -2,7 +2,8 @@
 
 ## Version 4.7.1
 - Bugfix :
-  - [PRE-2328](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2367): Fix payment method in database
+  - [SMP-2328](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2367): Fix payment method in database
+  - [PRE-2222](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2368): Fix order detail display
 
 ## Version 4.7.0
 - Refactoring :